:beetle: Fix tests.
