+++ frontend/src/__tests__/components/Sidebar/DistrictMap.test.tsx
@@ -5,13 +5,13 @@ import React from 'react';
 import {describe, test, vi} from 'vitest';
 import {render, screen} from '@testing-library/react';
 
-import i18n from '../../../../util/i18nForTests';
+import i18n from '../../../util/i18nForTests';
 
 import {I18nextProvider} from 'react-i18next';
 import {Provider} from 'react-redux';
-import {Store} from '../../../../store';
-import DistrictMap from '../../../../components/Sidebar/DistrictMap';
-import Theme from '../../../../util/Theme';
+import {Store} from '../../../store';
+import DistrictMap from '../../../components/Sidebar/DistrictMap';
+import Theme from '../../../util/Theme';
 import {ThemeProvider} from '@mui/material/styles';
 
 describe('District Map', () => {
