+++ frontend/src/__tests__/components/Sidebar/SearchBar.test.tsx
@@ -6,13 +6,13 @@ import {describe, test, expect, vi, afterEach} from 'vitest';
 import {act, cleanup, render, screen, waitFor} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import i18n from '../../../../util/i18nForTests';
+import i18n from '../../../util/i18nForTests';
 
-import SearchBar from '../../../../components/Sidebar/SearchBar';
+import SearchBar from '../../../components/Sidebar/SearchBar';
 import {I18nextProvider} from 'react-i18next';
 import {Provider} from 'react-redux';
-import {Store} from '../../../../store';
-import {selectDistrict} from '../../../../store/DataSelectionSlice';
+import {Store} from '../../../store';
+import {selectDistrict} from '../../../store/DataSelectionSlice';
 
 describe('SearchBar', () => {
   // We mock fetch to return two entries for searchable districts.
