+++ frontend/src/__tests__/components/Sidebar/SidebarHistory.test.tsx
@@ -5,9 +5,9 @@ import React from 'react';
 import {describe, test} from 'vitest';
 import {render} from '@testing-library/react';
 
-import i18n from '../../../../util/i18nForTests';
+import i18n from '../../../util/i18nForTests';
 
-import SidebarHistory from '../../../../components/Sidebar/SidebarHistory';
+import SidebarHistory from '../../../components/Sidebar/SidebarHistory';
 import {I18nextProvider} from 'react-i18next';
 
 describe('History', () => {
