+++ frontend/src/__tests__/components/TopBar/PopUps/AccessibilityDialog.test.tsx
@@ -6,12 +6,11 @@ import {describe, test} from 'vitest';
 import {render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import i18n from '../../../../../util/i18nForTests';
-
+import i18n from '../../../../util/i18nForTests';
 import {I18nextProvider} from 'react-i18next';
-import ApplicationMenu from '../../../../../components/TopBar/ApplicationMenu';
+import ApplicationMenu from '../../../../components/TopBar/ApplicationMenu';
 
-describe('ImprintDialog', () => {
+describe('AccessibilityDialog', () => {
   test('PopUp', async () => {
     render(
       <I18nextProvider i18n={i18n}>
@@ -23,10 +22,10 @@ describe('ImprintDialog', () => {
     const menu = screen.getByLabelText('topBar.menu.label');
     await userEvent.click(menu);
 
-    const imprint = screen.getByText('topBar.menu.imprint');
-    await userEvent.click(imprint);
+    const a11y = screen.getByText('topBar.menu.accessibility');
+    await userEvent.click(a11y);
 
-    await screen.findByText('imprint.header');
-    await screen.findByText('imprint.content');
+    await screen.findByText('accessibility.header');
+    await screen.findByText('accessibility.content');
   });
 });
