+++ frontend/src/__tests__/components/TopBar/PopUps/AttributionDialog.test.tsx
@@ -6,11 +6,11 @@ import {describe, test, vi, expect} from 'vitest';
 import {act, render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import i18n from '../../../../../util/i18nForTests';
+import i18n from '../../../../util/i18nForTests';
 
 import {I18nextProvider} from 'react-i18next';
 import {forceVisible} from 'react-lazyload';
-import ApplicationMenu from '../../../../../components/TopBar/ApplicationMenu';
+import ApplicationMenu from '../../../../components/TopBar/ApplicationMenu';
 
 describe('AttributionDialog', () => {
   test('PopUp', async () => {
