+++ frontend/src/__tests__/components/TopBar/PopUps/ImprintDialog.test.tsx
@@ -6,11 +6,12 @@ import {describe, test} from 'vitest';
 import {render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import i18n from '../../../../../util/i18nForTests';
-import {I18nextProvider} from 'react-i18next';
-import ApplicationMenu from '../../../../../components/TopBar/ApplicationMenu';
+import i18n from '../../../../util/i18nForTests';
 
-describe('PrivacyPolicyDialog', () => {
+import {I18nextProvider} from 'react-i18next';
+import ApplicationMenu from '../../../../components/TopBar/ApplicationMenu';
+
+describe('ImprintDialog', () => {
   test('PopUp', async () => {
     render(
       <I18nextProvider i18n={i18n}>
@@ -22,10 +23,10 @@ describe('PrivacyPolicyDialog', () => {
     const menu = screen.getByLabelText('topBar.menu.label');
     await userEvent.click(menu);
 
-    const pp = screen.getByText('topBar.menu.privacy-policy');
-    await userEvent.click(pp);
+    const imprint = screen.getByText('topBar.menu.imprint');
+    await userEvent.click(imprint);
 
-    await screen.findByText('privacy-policy.header');
-    await screen.findByText('privacy-policy.content');
+    await screen.findByText('imprint.header');
+    await screen.findByText('imprint.content');
   });
 });
