+++ frontend/src/__tests__/components/TopBar/PopUps/PrivacyPolicyDialog.test.tsx
@@ -6,11 +6,11 @@ import {describe, test} from 'vitest';
 import {render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import i18n from '../../../../../util/i18nForTests';
+import i18n from '../../../../util/i18nForTests';
 import {I18nextProvider} from 'react-i18next';
-import ApplicationMenu from '../../../../../components/TopBar/ApplicationMenu';
+import ApplicationMenu from '../../../../components/TopBar/ApplicationMenu';
 
-describe('AccessibilityDialog', () => {
+describe('PrivacyPolicyDialog', () => {
   test('PopUp', async () => {
     render(
       <I18nextProvider i18n={i18n}>
@@ -22,10 +22,10 @@ describe('AccessibilityDialog', () => {
     const menu = screen.getByLabelText('topBar.menu.label');
     await userEvent.click(menu);
 
-    const a11y = screen.getByText('topBar.menu.accessibility');
-    await userEvent.click(a11y);
+    const pp = screen.getByText('topBar.menu.privacy-policy');
+    await userEvent.click(pp);
 
-    await screen.findByText('accessibility.header');
-    await screen.findByText('accessibility.content');
+    await screen.findByText('privacy-policy.header');
+    await screen.findByText('privacy-policy.content');
   });
 });
