+++ frontend/src/__tests__/components/TopBar/TopBar.test.tsx
@@ -5,9 +5,9 @@ import React from 'react';
 import {describe, test} from 'vitest';
 import {render, screen} from '@testing-library/react';
 
-import i18n from '../../../../util/i18nForTests';
+import i18n from '../../../util/i18nForTests';
 
-import TopBar from '../../../../components/TopBar';
+import TopBar from '../../../components/TopBar';
 import {I18nextProvider} from 'react-i18next';
 
 describe('TopBar', () => {
