+++ frontend/src/__tests__/util/hooks.test.ts
@@ -4,7 +4,7 @@
 import {describe, test, expect} from 'vitest';
 import {renderHook} from '@testing-library/react';
 
-import {NumberFormatter} from '../../../util/hooks';
+import {NumberFormatter} from '../../util/hooks';
 
 describe('NumberFormatter', () => {
   test('languages', () => {
