+++ frontend/src/__tests__/util/util.test.ts
@@ -3,8 +3,8 @@
 
 import {describe, test, expect} from 'vitest';
 
-import {dateToISOString} from '../../../util/util';
-import * as util from '../../../util/util';
+import {dateToISOString} from '../../util/util';
+import * as util from '../../util/util';
 
 describe('deepCopy', () => {
   test('object', () => {
